Middleware ornatildi bu routelarga bormasdan oldin bizga jwt ni validate qilishni imkonini berad. , updatelar stockQuantity float tanlangan ekan bu  hato edi, intga ozgaritirldi, va price dayam hatolik boredi, int64 otqazilda tiyin da aniq korsatdi, docker compose up, mainda jwt config qoshildi. Hozir faqat architecture scheloton tayorlangyapti hali bussiness logic boshlanmadi

+++ pkg/models/Product.go
@@ -1,8 +1,8 @@
 package models
 
 type Product struct {
-	ID            int64   `json:"id"`
-	Name          string  `json:"name"`
-	Price         int     `json:"price"`
-	StockQuantity float64 `json:"stock_quantity"`
+	ID            int64  `json:"id"`
+	Name          string `json:"name"`
+	Price         int64  `json:"price"`
+	StockQuantity int    `json:"stock_quantity"`
 }
